sdcard_image bbclass: more fixes
* parametrize vars
* move error check to the top
* reuse image when possible

+++ classes/sdcard_image.bbclass
@@ -1,23 +1,43 @@
 inherit image
 
-IMAGE_FSTYPES_append = " sdimg"
+# Add the fstypes we need
+IMAGE_FSTYPES_append = " tar.bz2 sdimg"
+
+# Change this to match your host distro
+LOSETUP ?= "/sbin/losetup"
+
+# Since these need to go in /etc/fstab we can hardcode them
+LOOPDEV ?= "/dev/loop1"
+LOOPDEV_BOOT ?= "/dev/loop2"
+LOOPDEV_FS ?= "/dev/loop3"
 
 IMAGE_CMD_sdimg () {
 	SDIMG=${WORKDIR}/sd.img
 
+	# sanity check fstab entry for boot partition mounting
+	if [ "x$(cat /etc/fstab | grep ${LOOPDEV_BOOT} | grep ${WORKDIR}/tmp-mnt-boot | grep user || true)" = "x" ]; then
+		echo "/etc/fstab entries need to be created with the user flag for the loop devices like:"
+		echo "${LOOPDEV_BOOT} ${WORKDIR}/tmp-mnt-boot msdos user 0 0"
+		echo "${LOOPDEV_FS} ${WORKDIR}/tmp-mnt-rootfs ext3 user 0 0"
+        false
+	fi
+
 	# cleanup loops
-	for loop in $(/sbin/losetup -a | grep ${SDIMG}); do
+	for loop in $(${LOSETUP} -a | grep ${SDIMG}); do
 		loop_dev=$(echo $loop|cut -d ":" -f 1)
 		umount $loop_dev || true
-		/sbin/losetup -d $loop_dev || true
+		${LOSETUP} -d $loop_dev || true
 	done
 
-	dd if=/dev/zero of=${SDIMG} bs=$(echo '255 * 63 * 512' | bc) count=444
-	LOOPDEV=$(/sbin/losetup -f)
-	/sbin/losetup ${LOOPDEV} ${SDIMG}
+	# If an SD image is already present, reuse and reformat it
+	if [ ! -e ${SDIMG} ] ; then
+		dd if=/dev/zero of=${SDIMG} bs=$(echo '255 * 63 * 512' | bc) count=444
+	fi
+
+	${LOSETUP} ${LOOPDEV} ${SDIMG}
 
 	# Create partition table
-	#dd if=/dev/zero of=${LOOPDEV} bs=1024 count=1024
+	dd if=/dev/zero of=${LOOPDEV} bs=1024 count=1024
 	SIZE=$(/sbin/fdisk -l ${LOOPDEV} | grep Disk | grep bytes | awk '{print $5}')
 	CYLINDERS=$(echo $SIZE/255/63/512 | bc)
 	{
@@ -35,13 +55,11 @@ IMAGE_CMD_sdimg () {
 	LOOPDEV_BLOCKS=$(/sbin/fdisk -l -u $LOOPDEV 2>&1 | grep FAT | perl -p -i -e "s/\s+/ /g"|cut -d " " -f 5)
 	LOOPDEV_BYTES=$(echo "$LOOPDEV_BLOCKS * 1024" | bc)
 
-	LOOPDEV_BOOT=$(/sbin/losetup -f)
-	/sbin/losetup ${LOOPDEV_BOOT} ${SDIMG} -o ${BOOT_OFFSET} 
+	${LOSETUP} -d ${LOOPDEV}
+
+	${LOSETUP} ${LOOPDEV_BOOT} ${SDIMG} -o ${BOOT_OFFSET} 
 
 	/sbin/mkfs.msdos ${LOOPDEV_BOOT} -n boot
-
-	LOOPDEV_FS=$(/sbin/losetup -f)
-	/sbin/losetup ${LOOPDEV_FS} ${SDIMG} -o ${FS_OFFSET}	
 
 	# Prepare filesystem partition
 	# Copy ubi used by flashing scripts
@@ -49,22 +67,19 @@ IMAGE_CMD_sdimg () {
 		echo "Copying UBIFS image to file system"
 		cp ${DEPLOY_DIR_IMAGE}/${IMAGE_NAME}.rootfs.ubi ${IMAGE_ROOTFS}/boot/fs.ubi
 	fi
-	ROOTFS_SIZE="$(du -ks ${IMAGE_ROOTFS} | awk '{print 65536 + $1}')"
-	genext2fs -b ${FS_SIZE_BLOCKS} -d ${IMAGE_ROOTFS} ${LOOPDEV_FS}
-	/sbin/tune2fs -j ${LOOPDEV_FS}
 
 	# Prepare boot partion. First mount the boot partition, and copy the boot loader and supporting files
 	# from the root filesystem
 
 	# sanity check fstab entry for boot partition mounting
-	if [ "x$(cat /etc/fstab | grep $LOOPDEV_BOOT | grep ${WORKDIR}/tmp-mnt-boot | grep user || true)" = "x" ]; then
+	if [ "x$(cat /etc/fstab | grep ${WORKDIR}/tmp-mnt-boot | grep user || true)" = "x" ]; then
 		echo "/etc/fstab entries need to be created with the user flag for $LOOPDEV_BOOT like:"
 		echo "$LOOPDEV_BOOT ${WORKDIR}/tmp-mnt-boot msdos user 0 0"
-		false
+		#false
 	fi
 
 	mkdir -p ${WORKDIR}/tmp-mnt-boot
-	mount $LOOPDEV_BOOT
+	mount $LOOPDEV_BOOT ${WORKDIR}/tmp-mnt-boot
 
 	echo "Copying bootloaders into the boot partition"
 	if [ -e ${IMAGE_ROOTFS}/boot/MLO ] ; then
@@ -86,11 +101,22 @@ IMAGE_CMD_sdimg () {
 		cp -v ${DEPLOY_DIR_IMAGE}/u-boot-${MACHINE}.$suffix ${WORKDIR}/tmp-mnt-boot 
 	fi
 
-	# cleanup
+
 	umount ${LOOPDEV_BOOT}
-	/sbin/losetup -d ${LOOPDEV}
-	/sbin/losetup -d ${LOOPDEV_BOOT}
-	/sbin/losetup -d ${LOOPDEV_FS}
+	${LOSETUP} -d ${LOOPDEV_BOOT}
+
+	# Ext3 section
+
+	ROOTFS_SIZE="$(du -ks ${IMAGE_ROOTFS} | awk '{print 65536 + $1}')"
+	${LOSETUP} ${LOOPDEV_FS} ${SDIMG} -o ${FS_OFFSET}
+	/sbin/mkfs.ext3 ${LOOPDEV_FS} -L ${IMAGE_NAME}
+
+	mkdir -p ${WORKDIR}/tmp-mnt-rootfs
+	mount $LOOPDEV_BOOT ${WORKDIR}/tmp-mnt-rootfs
+
+ 	tar jxf ${DEPLOY_DIR_IMAGE}/${IMAGE_NAME}.rootfs.tar.bz2 -C ${WORKDIR}/tmp-mnt-rootfs
+
+	${LOSETUP} -d ${LOOPDEV_FS}
 
 	gzip -c ${WORKDIR}/sd.img > ${DEPLOY_DIR_IMAGE}/${IMAGE_NAME}-${PR}.img.gz
 }
